Refactoring: Do away with constexpr and breakout headeronly framebuffer file
